git-svn-id: https://svn.apache.org/repos/asf/incubator/jena/Import/Jena-SVN/SDB/trunk@1115590 13f79535-47bb-0310-9956-ffa450edef68

+++ src/com/hp/hpl/jena/sdb/graph/GraphSDB.java
@@ -180,7 +180,7 @@ public class GraphSDB extends GraphBase implements Graph
     public TransactionHandler getTransactionHandler() { return store.getConnection().getTransactionHandler() ; }
     
     @Override
-    public int graphBaseSize() { return store.getSize(); }
+    public int graphBaseSize() { return (int)store.getSize(); }
 }
 
 /*
